Release 2.0.26+2007.4

+++ script/minimize.sh
@@ -1,8 +1,5 @@
 #!/bin/bash -eux
 
-if [[ "$DESKTOP" =~ ^(true|yes|on|1|TRUE|YES|ON])$ ]]; then
-  exit
-fi
 
 echo "==> Disk usage before minimization"
 df -h
@@ -46,12 +43,15 @@ while [ -n "$(deborphan --guess-all --libdevel)" ]; do
 done
 apt-get -y purge deborphan dialog
 
-echo "==> Removing man pages"
-rm -rf /usr/share/man/*
+if [[ "$CLEANUP_DOCS" =~ ^(true|yes|on|1|TRUE|YES|ON)$ ]]; then
+  echo "==> Removing man pages"
+  rm -rf /usr/share/man/*
+  echo "==> Removing any docs"
+  rm -rf /usr/share/doc/*
+fi
+
 echo "==> Removing APT files"
 find /var/lib/apt -type f | xargs rm -f
-echo "==> Removing any docs"
-rm -rf /usr/share/doc/*
 echo "==> Removing caches"
 find /var/cache -type f -exec rm -rf {} \;
 # delete any logs that have built up during the install
